manual: Change structure & Refactor to make code up to date
